@@ -1,7 +1,8 @@
-from pydantic_ome_ngff.v04 import Axis as AxisV04
+import pydantic_ome_ngff.v04.axis as axisV04
 
+AxisType = axisV04.AxisType
 
-class Axis(AxisV04):
+class Axis(axisV04.Axis):
     """
     Axis metadata.
     See https://ngff.openmicroscopy.org/latest/#axes-md
