@@ -1,4 +1,7 @@
 from __future__ import annotations
+from typing import Annotated
+
+from pydantic import AfterValidator, model_validator
 
 from pydantic_ome_ngff.base import StrictVersionedBase
 from pydantic_ome_ngff.v04.base import version
@@ -67,41 +70,34 @@ class TimeUnit(str, Enum):
     zettasecond = "zettasecond"
 
 
-class Axis(StrictVersionedBase):
+def check_type_unit(model: Axis) -> "Axis":
     """
-    Axis metadata.
-    See https://ngff.openmicroscopy.org/0.4/#axes-md
+    Check that the `unit` attribute of an `Axis` object is valid. This function emits warnings when the units of an `Axis` object
+    are spec-compliant but go against a "SHOULD" statement in the spec.
     """
 
-    _version = version
-    # SPEC: this should almost certainly be a string, but the spec doesn't specify the type: https://github.com/ome/ngff/blob/ee4d5dab677636a28f1f65c248a751e279a0d1fe/0.4/index.bs#L243
-    name: str
-    type: str | None
-    unit: str | None
-
-
-def check_unit(model: Axis) -> "Axis":
     typ = model.type
     unit = model.unit
+
     if typ == AxisType.space:
-        if unit not in SpaceUnit.__members__:
+        if unit not in [e.value for e in SpaceUnit]:
             msg = (
                 f"Unit '{unit}' is not recognized as a standard unit "
                 f"for an axis with type '{typ}'."
             )
             warnings.warn(msg)
     elif typ == AxisType.time:
-        if unit not in TimeUnit.__members__:
+        if unit not in [e.value for e in TimeUnit]:
             msg = (
-                f"Unit '{unit}' is not recognized as a standard unit for an axis "
-                f"with type '{typ}'."
+                f"Unit '{unit}' is not recognized as a standard unit "
+                f"for an axis with type '{typ}'."
             )
             warnings.warn(msg)
     elif typ == AxisType.channel:
         pass
     elif typ is None:
         msg = (
-            f"Null axis type. Version {model._version} of the OME-NGFF spec states "
+            f"The `type` field of this axis was set to `None`. Version {model._version} of the OME-NGFF spec states "
             "that the 'type' field of an axis should be set to a string."
         )
         warnings.warn(msg, UserWarning)
@@ -115,8 +111,26 @@ def check_unit(model: Axis) -> "Axis":
 
     if unit is None:
         msg = (
-            f"Null unit. Version {model._version} of the OME-NGFF spec states "
+            f"The `unit` field of this axis was set to `None`. Version {model._version} of the OME-NGFF spec states "
             "that the `unit` field of an axis should be set to a string."
         )
         warnings.warn(msg)
     return model
+
+
+
+class Axis(StrictVersionedBase):
+    """
+    Axis metadata.
+    See https://ngff.openmicroscopy.org/0.4/#axes-md
+    """
+
+    _version = version
+    # SPEC: this should almost certainly be a string, but the spec doesn't specify the type: https://github.com/ome/ngff/blob/ee4d5dab677636a28f1f65c248a751e279a0d1fe/0.4/index.bs#L243
+    name: str
+    type: str | None
+    unit: str | None
+
+    @model_validator(mode='after')
+    def check_type_unit(self):
+        return check_type_unit(self)
